#235: devtools page lifecycle when navigating
