+++ src/devTools/context.ts
@@ -22,13 +22,15 @@ import {
   detectFrameworks,
   getTabInfo,
   injectScript,
-  awaitPermissions,
-  waitNavigation,
-} from "@/background/devtools";
+  navigationEvent,
+  waitReady,
+} from "@/background/devtools/index";
 import useAsyncEffect from "use-async-effect";
 import { useAsyncState } from "@/hooks/common";
 import { FrameworkMeta } from "@/messaging/constants";
 import { reportError } from "@/telemetry/logging";
+import { v4 as uuidv4 } from "uuid";
+import { useTabEventListener } from "@/hooks/events";
 
 interface FrameMeta {
   url: string;
@@ -39,9 +41,9 @@ interface FrameConnectionState {
   frameId: number;
 
   /**
-   * Sequence number for the frameId
+   * UUID for the navigation result
    */
-  navSequence: number;
+  navSequence: string | undefined;
 
   /**
    * True if the devtools have permission to access the current tab
@@ -57,7 +59,7 @@ interface FrameConnectionState {
 }
 
 const initialFrameState: Omit<FrameConnectionState, "frameId"> = {
-  navSequence: 0,
+  navSequence: undefined,
   hasPermissions: false,
   error: undefined,
   meta: undefined,
@@ -68,6 +70,11 @@ export interface Context {
    * Re-connect to the background page
    */
   connect: () => Promise<void>;
+
+  /**
+   * True if the a connection attempt is in process
+   */
+  connecting: boolean;
 
   /**
    * The background page port.
@@ -84,6 +91,7 @@ export interface Context {
 
 const initialValue: Context = {
   connect: () => Promise.resolve(),
+  connecting: false,
   port: null,
   portError: undefined,
   tabState: { ...initialFrameState, frameId: 0 },
@@ -91,36 +99,36 @@ const initialValue: Context = {
 
 export const DevToolsContext = React.createContext(initialValue);
 
-/**
- * Returns a monotonically increasing number whenever the inspected page has a navigation event.
- */
-export function useNavSequence(port: Runtime.Port): number {
-  const tabId = browser.devtools.inspectedWindow.tabId;
-  const [seq, setSeq] = useState(0);
-
-  useAsyncEffect(
-    async (isMounted) => {
-      if (port) {
-        const details = await waitNavigation(port, tabId);
-        console.debug(`Target page navigated: ${details?.url}`, { details });
-        if (!isMounted()) {
-          return;
-        }
-        setSeq((prevState) => prevState + 1);
-      }
-    },
-    [tabId, seq, setSeq, port]
-  );
-
-  return seq;
-}
-
 class PermissionsError extends Error {
   constructor(msg: string) {
     super(msg);
     // Set the prototype explicitly.
     Object.setPrototypeOf(this, PermissionsError.prototype);
   }
+}
+
+function runInMillis<TResult>(
+  factory: () => Promise<TResult>,
+  maxMillis: number
+): Promise<TResult> {
+  return new Promise<TResult>((resolve, reject) => {
+    let settled = false;
+
+    const intervalId = setInterval(() => {
+      if (!settled) {
+        settled = true;
+        reject(`Method did not complete in ${maxMillis}ms`);
+      }
+    }, maxMillis);
+
+    factory()
+      .then(resolve)
+      .catch(reject)
+      .finally(() => {
+        settled = true;
+        clearInterval(intervalId);
+      });
+  });
 }
 
 async function connectToFrame(port: Runtime.Port): Promise<FrameMeta> {
@@ -131,78 +139,89 @@ async function connectToFrame(port: Runtime.Port): Promise<FrameMeta> {
   }
   console.debug(`connectToFrame: ensuring contentScript for ${url}`);
   await injectScript(port, { file: "contentScript.js" });
-  console.debug(`connectToFrame: detecting frameworks on ${url}`);
-  const frameworks = await detectFrameworks(port);
+
+  console.debug(
+    `connectToFrame: waiting for contentScript to be ready for ${url}`
+  );
+  await waitReady(port, { maxWaitMillis: 4000 });
+
+  let frameworks: FrameworkMeta[] = [];
+  try {
+    console.debug(`connectToFrame: detecting frameworks on ${url}`);
+    frameworks = await runInMillis(() => detectFrameworks(port), 500);
+  } catch (error) {
+    console.debug(`connectToFrame: error detecting frameworks ${url}`, {
+      error,
+    });
+  }
+
   console.debug(`connectToFrame: finished for ${url}`);
   return { frameworks, url };
 }
 
 export function useDevConnection(): Context {
+  const tabId = browser.devtools.inspectedWindow.tabId;
+
+  const [connecting, setConnecting] = useState(false);
+
   const [port, , portError] = useAsyncState(connectDevtools, []);
-  const navSequence = useNavSequence(port);
+
   const [current, setCurrent] = useState<FrameConnectionState>({
     ...initialFrameState,
     frameId: 0,
-    navSequence,
   });
-  // connectSeq is used to fake recursive calls to connect
-  const [connectSeq, setConnectSeq] = useState(0);
 
-  const connect = useCallback(
-    async (nav: number) => {
-      if (!port) {
-        throw new Error("background port not initialized");
-      }
-      const common = { frameId: 0, navSequence: nav };
-      try {
-        console.debug(`useDevConnection.connect: connecting for ${nav}`);
-        const meta = await connectToFrame(port);
-        console.debug(
-          `useDevConnection.connect: replacing tabState for ${nav}: ${meta.url}`
-        );
+  const connect = useCallback(async () => {
+    if (!port) {
+      throw new Error("background port not initialized");
+    }
+    const uuid = uuidv4();
+    const common = { frameId: 0, navSequence: uuid };
+    try {
+      console.debug(`useDevConnection.connect: connecting for ${uuid}`);
+      setConnecting(true);
+      const meta = await connectToFrame(port);
+      console.debug(
+        `useDevConnection.connect: replacing tabState for ${uuid}: ${meta.url}`
+      );
+      setCurrent({
+        ...common,
+        hasPermissions: true,
+        meta,
+      });
+    } catch (err) {
+      if (err instanceof PermissionsError) {
+        setCurrent({
+          ...common,
+          hasPermissions: false,
+          meta: undefined,
+        });
+      } else {
+        reportError(err);
         setCurrent({
           ...common,
           hasPermissions: true,
-          meta,
+          meta: undefined,
+          error: err.message?.toString() ?? err.toString(),
         });
-      } catch (err) {
-        if (err instanceof PermissionsError) {
-          setCurrent({
-            ...common,
-            hasPermissions: false,
-            meta: undefined,
-          });
-          awaitPermissions(port).then(() => {
-            setConnectSeq((prev) => prev + 1);
-          });
-        } else {
-          reportError(err);
-          setCurrent({
-            ...common,
-            hasPermissions: true,
-            meta: undefined,
-            error: err.message?.toString() ?? err.toString(),
-          });
-        }
       }
-    },
-    [port, setCurrent, setConnectSeq]
-  );
-
-  const externalConnect = useCallback(async () => {
-    await connect(navSequence);
-  }, [connect, navSequence]);
+    }
+    setConnecting(false);
+  }, [port, setCurrent]);
 
   // automatically connect on when background port connected, and on future navigations
   useAsyncEffect(async () => {
     if (port) {
-      await connect(navSequence);
+      await connect();
     }
-  }, [connectSeq, navSequence, port]);
+  }, [port]);
+
+  useTabEventListener(tabId, navigationEvent, connect);
 
   return {
     port,
-    connect: externalConnect,
+    connecting,
+    connect: connect,
     portError: portError?.toString(),
     tabState: current,
   };
