+++ src/vendors/libraryDetector/detect.ts
@@ -2,7 +2,7 @@
 // https://github.com/johnmichel/Library-Detector-for-Chrome/blob/4c77e4e26197763ce87ea878be047c9c1d6f7017/content_scripts/lib_detect.js
 
 // @ts-ignore
-import tests from "./libraries";
+import libraryTests from "./libraries";
 import { zip } from "lodash";
 
 interface Library {
@@ -10,28 +10,36 @@ interface Library {
   version: string | null;
 }
 
+type TestResult = { version: string | null } | false;
+
 interface Test {
   id: string;
-  test: (win: Window) => { version: string | null };
+  test: (win: Window) => TestResult;
 }
 
 async function detectLibraries(): Promise<Library[]> {
-  const libraries: Library[] = [];
-  const testDefinitions = Object.values(tests) as Test[];
+  const detectedLibraries: Library[] = [];
+  const tests = Object.values(libraryTests) as Test[];
   const results = await Promise.allSettled(
-    testDefinitions.map((x) => x.test(window))
+    tests.map(
+      (test) => new Promise<TestResult>((resolve) => resolve(test.test(window)))
+    )
   );
-  for (const [definition, result] of zip(testDefinitions, results)) {
+  for (const [definition, result] of zip(tests, results)) {
     if (result.status === "fulfilled") {
-      libraries.push({
-        id: definition.id,
-        version: result.value.version,
-      });
+      if (result.value !== false) {
+        detectedLibraries.push({
+          id: definition.id,
+          version: result.value.version,
+        });
+      }
     } else {
-      console.warn(`Library Detector test for ${definition.id} failed:`);
+      console.warn(
+        `Library Detector test for ${definition.id} failed: ${result.status}`
+      );
     }
   }
-  return libraries;
+  return detectedLibraries;
 }
 
 export default detectLibraries;
